数据库信息创建

+++ memory-voice-admin/src/main/java/com/mem/vo/business/base/model/po/VisitComment.java
@@ -16,7 +16,7 @@ public class VisitComment {
     /**
      * 访谈id
      */
-    private Integer visitId;
+    private long visitId;
 
     /**
      * 用户昵称
