+++ memory-voice-admin/src/main/java/com/mem/vo/business/base/model/po/VisitLike.java
@@ -16,7 +16,7 @@ public class VisitLike {
     /**
      * 访谈id
      */
-    private Integer visitId;
+    private long visitId;
 
     /**
      * 用户昵称
